requirements.txt
numpy==1.22.4
Pillow==9.3.0
tqdm==4.64.1
six==1.16.0
ultralytics==8.0.118